feat: moved augment into core
Created testutil and moved augmentors to core/ui instead of aug bc they
are specific to ui augmentation.
