+++ core/run/cmd.go
@@ -1,4 +1,4 @@
-package augment
+package run
 
 import (
 	"fmt"
@@ -7,10 +7,10 @@ import (
 )
 
 // Runs an augmentation command.
-// Gives you the option to customize how you want to log the command.
-func RunCommand(cmd *exec.Cmd, logger io.Writer) error {
-	cmd.Stderr = logger
-	cmd.Stdout = logger
+// Gives you the option to customize how you want to log the command output.
+func Cmd(cmd *exec.Cmd, writer io.Writer) error {
+	cmd.Stderr = writer
+	cmd.Stdout = writer
 	err := cmd.Run() //blocks until sub process is complete
 	if err != nil {
 		return fmt.Errorf("'%s' failed with err: %s", cmd.String(), err.Error())
